# swagger.yaml
openapi: 3.0.0
paths:
...
  /auth/login:
    post:
      operationId: AuthController_login
      summary: Login user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthDto'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                    example: >-
                      eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJyb2xlIjoiQURNSU4iLCJzdWIiOiIxIiwiaWF0IjoxNzE5ODYyNTQzLCJleHAiOjE3MTk4NjYxNDN9.qMOYT-hipMALLcK3kHIkMFaxcQQLH4bhbtkMl50-t-c
                    description: JWT token
        '400':
          description: Bad Request - Caused by an incorrect request object
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Email must be a maximum of 50 characters
                        - Email is not valid
                        - Email cannot be empty
                        - Email must be a string
                      description: List of errors
                  error:
                    type: string
                    example: BadRequestException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.421Z'
                    description: Date of the error
        '401':
          description: Unauthorized - Caused by an invalid user or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Data not valid
                      description: List of errors
                  error:
                    type: string
                    example: UnauthorizedException
                    description: Error type
                  status:
                    type: number
                    example: 401
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.421Z'
                    description: Date of the error
      tags: &ref_0
        - Auth
  /auth/logout:
    post:
      operationId: AuthController_logout
      summary: Logout user
      parameters: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Logout successful
                    description: Logout message
        '401':
          description: Unauthorized - Caused by an invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Invalid token
                      description: List of errors
                  error:
                    type: string
                    example: InvalidTokenException
                    description: Error type
                  status:
                    type: number
                    example: 401
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.421Z'
                    description: Date of the error
      tags: *ref_0
      security:
        - access-token: []
  /histories/earnings/parkingId/{parkingId}:
    get:
      operationId: HistoriesController_getEarnings
      summary: Get earnings from a specific parking lot
      parameters:
        - name: parkingId
          required: true
          in: path
          description: Parking ID
          schema:
            type: number
      responses:
        '200':
          description: Earnings obtained successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EarningsResponseDto'
        '401':
          description: Unauthorized - Caused by Invalid token
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Invalid token
                      description: List of errors
                  error:
                    type: string
                    example: InvalidTokenException
                    description: Error type
                  status:
                    type: number
                    example: 401
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.386Z'
                    description: Date of the error
      tags:
        - Histories
      security:
        - access-token: []
...
  /sending-mail:
    post:
      operationId: MailController_sendMail
      summary: Send mail
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MailRequestDto'
      responses:
        '200':
          description: Mail sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MailResponseDto'
        '400':
          description: Bad Request - Caused by an incorrect request object
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Email must be a maximum of 50 characters
                        - Email is not valid
                        - Email cannot be empty
                        - Email must be a string
                      description: List of errors
                  error:
                    type: string
                    example: BadRequestException
                    description: Error type
                  status:
                    type: number
                    example: 400
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.452Z'
                    description: Date of the error
        '401':
          description: >-
            Unauthorized - Caused by user does not have permission to access
            this route
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - User does not have permission to access this route
                      description: List of errors
                  error:
                    type: string
                    example: JwtAuthException
                    description: Error type
                  status:
                    type: number
                    example: 401
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.452Z'
                    description: Date of the error
        '500':
          description: Error - Internal server error occurred
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: array
                    items:
                      type: string
                      example:
                        - Error sending mail
                      description: List of errors
                  error:
                    type: string
                    example: InternalServerErrorException
                    description: Error type
                  status:
                    type: number
                    example: 500
                    description: HTTP status code
                  date:
                    type: string
                    example: '2024-07-09T22:07:00.452Z'
                    description: Date of the error
      tags:
        - Mail
      security:
        - access-token: []
info:
  title: PARKING CRUD API
  description: >-
    The Rest API manages vehicle parking, allows partner creation for vehicle
    entry and exit, and uses JWT tokens with role-based access control (ADMIN
    and SOCIO) expiring every 6 hours, built with NestJS and PostgreSQL
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    access-token:
      scheme: bearer
      bearerFormat: JWT
      type: http
      in: header
  schemas:
...
    AuthDto:
      type: object
      properties:
        email:
          type: string
          description: User email
          maxLength: 50
          example: admin@mail.com
        password:
          type: string
          description: User password
          example: admin
      required:
        - email
        - password
    EarningsResponseDto:
      type: object
      properties:
        today:
          type: number
          description: Today earnings
          example: 0
        week:
          type: number
          description: Week earnings
          example: 0
        month:
          type: number
          description: Month earnings
          example: 0
        year:
          type: number
          description: Year earnings
          example: 0
      required:
        - today
        - week
        - month
        - year
...
    MailRequestDto:
      type: object
      properties:
        email:
          type: string
          description: User email
          maxLength: 50
          example: socio@mail.com
        plate:
          type: string
          description: Vehicle plate
          example: ABC123
          minLength: 6
          maxLength: 6
        message:
          type: string
          description: Message to send
          example: Your parking time is about to expire
        parkingId:
          type: number
          description: Parking id
          example: 1
      required:
        - email
        - plate
        - message
        - parkingId
    MailResponseDto:
      type: object
      properties:
        message:
          type: string
          description: Message sent by the server
          example: Mail sent successfully
      required:
        - message